% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-objects.R
\docType{data}
\name{oce_example_data}
\alias{oce_example_data}
\title{A dataframe constructed for testing of OCEcharts.}
\format{A data frame with 20 rows and 4 variables}
\source{
\url{http://www.diamondse.info/}
}
\usage{
oce_example_data
}
\description{
A dataframe constructed for testing of OCEcharts.
}
\keyword{datasets}
