% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-objects.R
\docType{data}
\name{oce_guide_color}
\alias{oce_guide_color}
\title{A list modifying the format of the color legend to be consistent with OCE branding.}
\format{A list specifying format of the color legend}
\usage{
oce_guide_color
}
\description{
A list modifying the format of the color legend to be consistent with OCE branding.
}
\keyword{datasets}
