% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-objects.R
\docType{data}
\name{oce_palette}
\alias{oce_palette}
\title{A vector of orded color codes consistent with the OCE branding.}
\format{A vector of color codes.}
\usage{
oce_palette
}
\description{
A vector of orded color codes consistent with the OCE branding.
}
\keyword{datasets}
