% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-functions.R
\name{theme_oce}
\alias{theme_oce}
\title{This applies an OCE consistent theme to a chart created with ggplot2.}
\usage{
theme_oce()
}
\value{
A ggplot object consistent with OCE branding .
}
\description{
This applies an OCE consistent theme to a chart created with ggplot2.
}
\examples{

library(magrittr)
library(ggplot2)
library(OCEcharts)

oce_example_data \%>\%
   ggplot(aes(x = name, y = value_01, fill = as.factor(year))) +
   geom_bar(position = position_dodge(width = .8),stat = "identity", width = 0.6 ) +
   theme_oce() +
   scale_fill_manual(values = oce_palette[1:5]) + ## Would like to automate palette
   labs(x = "Category", y = "Per cent") +
   scale_y_continuous(limits = c(0,1), expand = c(0,0), breaks = seq(0,1,.2)) +
   oce_guide_fill

}
